pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shipwreck"
version = "0.1.0"
description = "Mapping the buried treasure in your container stack."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12",
    "pydantic>=2",
    "jinja2>=3",
    "pyyaml>=6",
    "python-hcl2>=4",
    "httpx>=0.27",
    "semver>=3",
    "rich>=13",
]

[project.scripts]
shipwreck = "shipwreck.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-mock>=3",
    "respx>=0.21",
    "pytest-snapshot>=0.9",
    "ruff>=0.4",
    "basedpyright>=1.18",
]

[tool.hatch.build.targets.wheel]
packages = ["src/shipwreck"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
line-ending = "lf"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.basedpyright]
pythonVersion = "3.12"
include = ["src"]
typeCheckingMode = "standard"
